+++ client/index.html
@@ -24,8 +24,8 @@
     -->
     <link rel="manifest" href="/manifest.json" />
     <link
-      rel="preload"
-      href="https://encore.scdn.co/fonts/CircularSp-Book-4eaffdf96f4c6f984686e93d5d9cb325.woff2"
+      rel="stylesheet"
+      href="https://fonts.googleapis.com/css2?family=Source+Sans+Pro:wght@400;700;900&display=swap"
       as="font"
       type="font/woff2"
       crossorigin="anonymous"
