+++ client/package.json
@@ -6,7 +6,7 @@
     "node": "18"
   },
   "dependencies": {
-    "@apollo/client": "^0.0.0-pr-10970-20230612181718",
+    "@apollo/client": "^3.8.0-beta.7",
     "@radix-ui/react-context-menu": "^2.1.3",
     "@radix-ui/react-dialog": "^1.0.3",
     "@radix-ui/react-dropdown-menu": "^2.0.4",
@@ -45,9 +45,8 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "codegen": "graphql-codegen",
     "start": "vite",
-    "build": "npm run codegen && tsc && vite build",
+    "build": "tsc && vite build",
     "serve": "serve dist -s -n -L -p $PORT",
     "lint": "eslint . --ext .ts,.tsx",
     "postinstall": "npm run build"
@@ -65,10 +64,7 @@
     ]
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "^4.0.1",
-    "@graphql-codegen/fragment-matcher": "^5.0.0",
-    "@graphql-codegen/typescript": "^4.0.0",
-    "@graphql-codegen/typescript-operations": "^4.0.0",
+    "@apollo/rover": "^0.16.2",
     "@mdx-js/rollup": "^2.3.0",
     "@tailwindcss/line-clamp": "^0.4.2",
     "@types/mdx": "^2.0.3",
