+++ client/src/components/CurrentUserMenu.tsx
@@ -1,7 +1,7 @@
-import { gql, useSuspenseQuery } from '@apollo/client';
-import { CurrentUserQuery, CurrentUserQueryVariables } from '../types/api';
-import DropdownMenu from './DropdownMenu';
-import Avatar from './Avatar';
+import { gql, useSuspenseQuery } from "@apollo/client";
+import { CurrentUserQuery, CurrentUserQueryVariables } from "../types/api";
+import DropdownMenu from "./DropdownMenu";
+import Avatar from "./Avatar";
 
 const CURRENT_USER_QUERY = gql`
   query CurrentUserQuery {
@@ -24,10 +24,15 @@ const CurrentUserMenu = () => {
   >(CURRENT_USER_QUERY);
 
   if (!data.me) {
-    throw new Error('You must be logged in');
+    throw new Error("You must be logged in");
   }
 
   const { user } = data.me;
+
+  const exploreSchemaUrl =
+    process.env.NODE_ENV == "production"
+      ? "https://studio.apollographql.com/public/spotify-ev3of9/variant/prod/home"
+      : `${import.meta.env.VITE_SERVER_HOST}/graphql`;
 
   return (
     <DropdownMenu>
@@ -41,7 +46,9 @@ const CurrentUserMenu = () => {
       </DropdownMenu.Trigger>
       <DropdownMenu.Menu align="end">
         <DropdownMenu.Item to="/settings">Settings</DropdownMenu.Item>
-        <DropdownMenu.Item to={`https://studio.apollographql.com/public/spotify-ev3of9/variant/prod/home`}>
+        <DropdownMenu.Item
+          to={exploreSchemaUrl}
+        >
           Explore the schema
         </DropdownMenu.Item>
         <DropdownMenu.Item to="https://github.com/apollographql/spotify-showcase">
