+++ client/src/router.tsx
@@ -25,17 +25,17 @@ import * as RootRoute from './routes/root';
 import * as SettingsRoute from './routes/settings';
 import * as ShowRoute from './routes/shows/show';
 import * as TrackRoute from './routes/tracks/track';
+import * as LoginRoute from './routes/login';
+import * as OAuthRoute from './routes/oauth/finalize';
 
 import RootErrorBoundary from './components/RootErrorBoundary';
-import { loader as loginLoader } from './routes/login';
-import { loader as oauthFinalizeLoader } from './routes/oauth/finalize';
 import { isLoggedInVar } from './vars';
 
 const routes = createRoutesFromElements(
   <Route path="/" errorElement={<RootErrorBoundary />}>
     <Route path="logout" loader={LogoutRoute.loader} />,
-    <Route path="login" loader={loginLoader} />,
-    <Route path="/oauth/finalize" errorElement={<RootErrorBoundary />} loader={oauthFinalizeLoader} />,
+    <Route path="login" loader={LoginRoute.loader} />,
+    <Route path="/oauth/finalize" errorElement={<RootErrorBoundary />} loader={OAuthRoute.loader} />,
     <Route
       path="logged-out"
       loader={LoggedOutRoute.loader}
