+++ subgraphs/playback/src/index.ts
@@ -176,7 +176,6 @@ async function main() {
     };
   };
 
-  app.use(express.static("public"));
   app.use(
     "/",
     cors(),
