--- subgraphs/playback/src/resolvers/PlaylistTrack.ts
@@ -1,12 +1,0 @@
-// import { PlaylistTrackResolvers } from '../__generated__/resolvers-types';
-
-// export const PlaylistTrack: PlaylistTrackResolvers = {
-//   __resolveType: (playlistItem) => {
-//     switch (playlistItem.type) {
-//       case 'track':
-//         return 'Track';
-//       case 'episode':
-//         return 'Episode';
-//     }
-//   },
-// };
