+++ subgraphs/playback/src/resolvers/helpers.ts
@@ -1,9 +1,9 @@
 import { OmitIndexSignature, ValueOf } from 'type-fest';
 import { GraphQLError } from 'graphql';
-// import { ContextValue } from '../types';
+import { ContextValue } from '../types/ContextValue';
 import { ResolversParentTypes, Resolver, ResolverFn } from '../__generated__/resolvers-types';
 import { getFieldConfig, identify } from '../fieldConfigs';
-// import { wait } from '../utils/common';
+import { wait } from '../utils/common';
 
 type ParentTypes = ValueOf<OmitIndexSignature<ResolversParentTypes>>;
 
@@ -13,30 +13,30 @@ export function prop<TParent extends ParentTypes, TKey extends keyof TParent>(
   return (parent: TParent) => parent[key];
 }
 
-// exp0ort function itself<TParent extends ParentTypes>(): Resolver<
-//   TParent,
-//   TParent
-// > {
-//   return (parent: TParent) => parent;
-// }
+export function itself<TParent extends ParentTypes>(): Resolver<
+  TParent,
+  TParent
+> {
+  return (parent: TParent) => parent;
+}
 
-// export function wrapWithSynthetics<TResult, TParent = {}, TArgs = {}>(
-//   resolver: ResolverFn<TResult, TParent, ContextValue, TArgs>
-// ): Resolver<TResult, TParent, ContextValue, TArgs> {
-//   return async (parent, args, context, info) => {
-//     const fieldConfig = getFieldConfig(identify.fromResolverInfo(info));
-//     const { config } = fieldConfig;
+export function wrapWithSynthetics<TResult, TParent = {}, TArgs = {}>(
+  resolver: ResolverFn<TResult, TParent, ContextValue, TArgs>
+): Resolver<TResult, TParent, ContextValue, TArgs> {
+  return async (parent, args, context, info) => {
+    const fieldConfig = getFieldConfig(identify.fromResolverInfo(info));
+    const { config } = fieldConfig;
 
-//     if (config.timeout > 0) {
-//       await wait(config.timeout);
-//     }
+    if (config.timeout > 0) {
+      await wait(config.timeout);
+    }
 
-//     if (Math.random() < config.errorRate) {
-//       throw new GraphQLError("You've been hit by a synthetic error!", {
-//         extensions: { code: 'SYNTHETIC_ERROR' },
-//       });
-//     }
+    if (Math.random() < config.errorRate) {
+      throw new GraphQLError("You've been hit by a synthetic error!", {
+        extensions: { code: 'SYNTHETIC_ERROR' },
+      });
+    }
 
-//     return resolver(parent, args, context, info);
-//   };
-// }
+    return resolver(parent, args, context, info);
+  };
+}
