--- subgraphs/spotify/src/@types/graphql.d.ts
@@ -1,6 +1,0 @@
-declare module '*.graphql' {
-  import { DocumentNode } from 'graphql';
-  const Schema: DocumentNode;
-
-  export = Schema;
-}
