+++ subgraphs/spotify/src/resolvers/helpers.ts
@@ -1,5 +1,9 @@
 import { OmitIndexSignature, ValueOf } from 'type-fest';
-import { ResolversParentTypes, Resolver } from '../__generated__/resolvers-types';
+import { ResolversParentTypes, Resolver, ResolverFn } from '../__generated__/resolvers-types';
+import { GraphQLError } from 'graphql';
+import { ContextValue } from '../types/ContextValue';
+import { getFieldConfig, identify } from '../fieldConfigs';
+import { wait } from '../utils/common';
 
 type ParentTypes = ValueOf<OmitIndexSignature<ResolversParentTypes>>;
 
@@ -14,4 +18,25 @@ export function itself<TParent extends ParentTypes>(): Resolver<
   TParent
 > {
   return (parent: TParent) => parent;
+}
+
+export function wrapWithSynthetics<TResult, TParent = {}, TArgs = {}>(
+  resolver: ResolverFn<TResult, TParent, ContextValue, TArgs>
+): Resolver<TResult, TParent, ContextValue, TArgs> {
+  return async (parent, args, context, info) => {
+    const fieldConfig = getFieldConfig(identify.fromResolverInfo(info));
+    const { config } = fieldConfig;
+
+    if (config.timeout > 0) {
+      await wait(config.timeout);
+    }
+
+    if (Math.random() < config.errorRate) {
+      throw new GraphQLError("You've been hit by a synthetic error!", {
+        extensions: { code: 'SYNTHETIC_ERROR' },
+      });
+    }
+
+    return resolver(parent, args, context, info);
+  };
 }
