+++ subgraphs/spotify/src/utils/common.ts
@@ -1,4 +1,7 @@
-export const isPlainObject = (obj: any): obj is object => {
+export const wait = (ms: number) =>
+  new Promise<void>((resolve) => setTimeout(resolve, ms));
+ 
+  export const isPlainObject = (obj: any): obj is object => {
   return typeof obj === 'object' && !Array.isArray(obj) && obj !== null;
 };
