+++ subgraphs/spotify/src/utils/constants.ts
@@ -16,8 +16,3 @@ export const OAUTH_SCOPES = [
   'user-top-read',
   'user-read-recently-played',
 ] as const;
-
-export const TOPICS = {
-  DISCONNECT: 'DISCONNECT',
-  PLAYBACK_STATE_CHANGED: 'PLAYBACK_STATE_CHANGED',
-} as const;
